LED PWM Phase 2
This second phase has two LEDs that alternate in brightness at the press of a button. When one LED brightens up, the other dims down. These represent the two servos tilting in a longitude direction. There also is a third LED that represents a servo sweeping laterally. The direction sets can be toggled with a switch.